+++ src/it/scala/com/exasol/cloudetl/DataImporterIT.scala
@@ -9,14 +9,12 @@ import java.nio.charset.StandardCharsets.UTF_8
 import java.sql.ResultSet
 import java.sql.Timestamp
 import java.time._
-
 import com.exasol.cloudetl.orc.OrcTestDataWriter
 import com.exasol.cloudetl.parquet.ParquetTestDataWriter
-import com.exasol.cloudetl.util.DateTimeUtil
+import com.exasol.cloudetl.helper.DateTimeHelper
 import com.exasol.matcher.CellMatcherFactory
 import com.exasol.matcher.ResultSetStructureMatcher.table
 import com.exasol.matcher.TypeMatchMode._
-
 import org.apache.avro.file.DataFileWriter
 import org.apache.avro._
 import org.apache.avro.generic._
@@ -630,11 +628,11 @@ class DataImporterIT extends BaseS3IntegrationTest {
         .withInputValues[Any](List(0, 1, 54, 567, 1234))
         .assertResultSet(
           table()
-            .row(DateTimeUtil.daysToDate(0))
-            .row(DateTimeUtil.daysToDate(1))
-            .row(DateTimeUtil.daysToDate(54))
-            .row(DateTimeUtil.daysToDate(567))
-            .row(DateTimeUtil.daysToDate(1234))
+            .row(DateTimeHelper.daysToDate(0))
+            .row(DateTimeHelper.daysToDate(1))
+            .row(DateTimeHelper.daysToDate(54))
+            .row(DateTimeHelper.daysToDate(567))
+            .row(DateTimeHelper.daysToDate(1234))
             .matches()
         )
     }
@@ -794,8 +792,8 @@ class DataImporterIT extends BaseS3IntegrationTest {
       val millis = System.currentTimeMillis()
       val timestamp = new Timestamp(millis)
       val buffer = ByteBuffer.allocate(12).order(ByteOrder.LITTLE_ENDIAN)
-      val micros = DateTimeUtil.getMicrosFromTimestamp(timestamp)
-      val (days, nanos) = DateTimeUtil.getJulianDayAndNanos(micros)
+      val micros = DateTimeHelper.getMicrosFromTimestamp(timestamp)
+      val (days, nanos) = DateTimeHelper.getJulianDayAndNanos(micros)
       buffer.putLong(nanos).putInt(days)
       val zdt = ZonedDateTime.ofInstant(Instant.ofEpochMilli(millis), ZoneId.of("Europe/Berlin"))
       val expectedTimestamp = Timestamp.valueOf(zdt.toLocalDateTime())
