+++ src/main/scala/com/exasol/cloudetl/helper/ParquetSchemaConverter.scala
@@ -1,4 +1,4 @@
-package com.exasol.cloudetl.util
+package com.exasol.cloudetl.helper
 
 import java.util.Locale.ENGLISH
