+++ src/test/scala/com/exasol/cloudetl/helper/ParquetSchemaConverterTest.scala
@@ -1,6 +1,7 @@
 package com.exasol.cloudetl.util
 
 import com.exasol.cloudetl.data.ExaColumnInfo
+import com.exasol.cloudetl.helper.ParquetSchemaConverter
 
 import org.apache.parquet.schema._
 import org.apache.parquet.schema.LogicalTypeAnnotation._
